backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-multipart
youtube_transcript_api
sentence-transformers
tqdm
numpy
requests
uvicorn
aiofiles
python-dotenv
pysrt
starlette
typing-extensions
langchain
pinecone
langchain-google-genai
rank_bm25